+++ sample/naruby/bench/compose.c
@@ -1,18 +1,19 @@
 #include <stdio.h>
+#include <stdint.h>
 
 #define A __attribute__((noinline,noipa))
 
-static A int h(int n) { return n + 1; }
-static A int g(int n) { return h(n) * 2; }
-static A int f(int n) { return g(n) - 3; }
+static A int64_t h(int64_t n) { return n + 1; }
+static A int64_t g(int64_t n) { return h(n) * 2; }
+static A int64_t f(int64_t n) { return g(n) - 3; }
 
 int main(void)
 {
-    int sz = 30 * 1000 * 1000;
-    int acc = 0;
-    for (int i = 0; i < sz; i++) {
+    int64_t sz = 30 * 1000 * 1000;
+    int64_t acc = 0;
+    for (int64_t i = 0; i < sz; i++) {
         acc += f(10);
     }
-    printf("%d\n", acc);
+    printf("%ld\n", (long)acc);
     return 0;
 }
