+++ sample/naruby/bench/loop.c
@@ -1,12 +1,13 @@
 #include <stdio.h>
+#include <stdint.h>
 
 int
 main(int argc, const char *argv[])
 {
-    int i;
+    int64_t i;
     for (i=0; i<1000*1000*100; i++) {
     }
-    printf("%d\n", i);
+    printf("%ld\n", (long)i);
     (void)argc; (void)argv;
     return 0;
 }
